Restructure runner
